[REF] web, calendar: use useModel instead of Record
This commit replaces the use of Record component by useModel hook, with
a customised RelationalModel that allows to create the model without
fetching the data, as we already have them on hand. Before that, we
have to make sure that the values given are in Unity format. This part
of code was previously in the Record component, but it is now moved to
this customised RelationalModel.

Part of task-id 3433922
