+++ addons/web/static/src/views/record.js
@@ -1,9 +1,7 @@
 /** @odoo-module **/
 
 import { useService } from "@web/core/utils/hooks";
-import { pick } from "@web/core/utils/objects";
 import { RelationalModel } from "@web/model/relational_model/relational_model";
-import { getFieldsSpec } from "@web/model/relational_model/utils";
 import { Component, xml, onWillStart, onWillUpdateProps, useState } from "@odoo/owl";
 
 const defaultActiveField = { attrs: {}, options: {}, domain: "[]", string: "" };
@@ -50,81 +48,16 @@ class _Record extends Component {
         modelServices.orm = this.orm;
         this.model = useState(new StandaloneRelationalModel(this.env, modelParams, modelServices));
 
-        const loadWithValues = async (values) => {
-            values = pick(values, ...Object.keys(modelParams.config.activeFields));
-            const proms = [];
-            for (const fieldName in values) {
-                if (["one2many", "many2many"].includes(this.props.fields[fieldName].type)) {
-                    if (values[fieldName].length && typeof values[fieldName][0] === "number") {
-                        const resModel = this.props.fields[fieldName].relation;
-                        const resIds = values[fieldName];
-                        const activeField = modelParams.config.activeFields[fieldName];
-                        if (activeField.related) {
-                            const { activeFields, fields } = activeField.related;
-                            const fieldSpec = getFieldsSpec(activeFields, fields, {});
-                            const kwargs = {
-                                context: activeField.context || {},
-                                specification: fieldSpec,
-                            };
-                            proms.push(
-                                this.orm.webRead(resModel, resIds, kwargs).then((records) => {
-                                    values[fieldName] = records;
-                                })
-                            );
-                        }
-                    }
-                }
-                if (this.props.fields[fieldName].type === "many2one") {
-                    const loadDisplayName = async (resId) => {
-                        const resModel = this.props.fields[fieldName].relation;
-                        const activeField = modelParams.config.activeFields[fieldName];
-                        const kwargs = {
-                            context: activeField.context || {},
-                            specification: { display_name: {} },
-                        };
-                        const records = await this.orm.webRead(resModel, [resId], kwargs);
-                        return records[0].display_name;
-                    };
-                    if (typeof values[fieldName] === "number") {
-                        const prom = loadDisplayName(values[fieldName]);
-                        prom.then((displayName) => {
-                            values[fieldName] = {
-                                id: values[fieldName],
-                                display_name: displayName,
-                            };
-                        });
-                        proms.push(prom);
-                    } else if (Array.isArray(values[fieldName])) {
-                        if (values[fieldName][1] === undefined) {
-                            const prom = loadDisplayName(values[fieldName][0]);
-                            prom.then((displayName) => {
-                                values[fieldName] = {
-                                    id: values[fieldName][0],
-                                    display_name: displayName,
-                                };
-                            });
-                            proms.push(prom);
-                        }
-                        values[fieldName] = {
-                            id: values[fieldName][0],
-                            display_name: values[fieldName][1],
-                        };
-                    }
-                }
-                await Promise.all(proms);
-            }
-            return this.model.load({ values });
-        };
         onWillStart(() => {
             if (this.props.values) {
-                return loadWithValues(this.props.values);
+                return this.model.load({ values: this.props.values });
             } else {
                 return this.model.load();
             }
         });
         onWillUpdateProps((nextProps) => {
             if (nextProps.values) {
-                return loadWithValues(nextProps.values);
+                return this.model.load({ values: nextProps.values });
             } else if (nextProps.info.resId !== this.model.root.resId) {
                 return this.model.load({ resId: nextProps.info.resId });
             }
